refactor: PhotoResponse를 Inner DTO class로 관리
