--- packy-api/src/main/java/com/dilly/gift/dto/response/PhotoResponse.java
@@ -1,20 +1,0 @@
-package com.dilly.gift.dto.response;
-
-import com.dilly.gift.domain.Photo;
-import lombok.Builder;
-
-@Builder
-public record PhotoResponse(
-    String photoUrl,
-    String description,
-    Integer sequence
-) {
-
-    public static PhotoResponse of(Photo photo) {
-        return PhotoResponse.builder()
-            .photoUrl(photo.getImgUrl())
-            .description(photo.getDescription())
-            .sequence(photo.getSequence())
-            .build();
-    }
-}
