+++ packy-api/src/test/java/com/dilly/gift/api/GiftBoxControllerTest.java
@@ -20,7 +20,7 @@ import com.dilly.gift.dto.response.EnvelopeResponse;
 import com.dilly.gift.dto.response.GiftBoxIdResponse;
 import com.dilly.gift.dto.response.GiftBoxResponse;
 import com.dilly.gift.dto.response.GiftResponse;
-import com.dilly.gift.dto.response.PhotoResponse;
+import com.dilly.gift.dto.response.PhotoResponseDto.PhotoResponse;
 import com.dilly.gift.dto.response.StickerResponse;
 import com.dilly.global.ControllerTestSupport;
 import com.dilly.global.WithCustomMockUser;
