+++ packy-api/src/test/java/com/dilly/gift/application/GiftBoxServiceTest.java
@@ -19,7 +19,7 @@ import com.dilly.gift.dto.response.BoxResponse;
 import com.dilly.gift.dto.response.GiftBoxIdResponse;
 import com.dilly.gift.dto.response.GiftBoxResponse;
 import com.dilly.gift.dto.response.GiftResponse;
-import com.dilly.gift.dto.response.PhotoResponse;
+import com.dilly.gift.dto.response.PhotoResponseDto.PhotoResponse;
 import com.dilly.gift.dto.response.StickerResponse;
 import com.dilly.global.IntegrationTestSupport;
 import com.dilly.global.WithCustomMockUser;
